move model to separate file
